pyproject.toml
[tool.poetry]
name = "s3-cli-tool"
version = "0.1.0"
description = ""
authors = ["kharshiladzemarita40 <kharshiladzemarita40@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.42.78"
python-dotenv = "^1.2.2"
python-magic = "^0.4.27"
typer = {extras = ["all"], version = "^0.24.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
s3-tool = "s3_cli_tool.main:app"